Create SimpleCalculatorApp.kt
This app allows the user to enter two numbers and perform basic arithmetic operations (addition, subtraction, multiplication, and division).
In this code, we define three instance variables for the EditText and TextView elements in our layout. In the onCreate method, we use the findViewById method to retrieve references to these elements. We also set up click listeners for the four buttons in our layout (btn_add, btn_subtract, btn_multiply, and btn_divide), and implement the logic to perform the arithmetic operations. Finally, we update the tvResult TextView to display the result of the calculation.

Note that this code assumes that you have a layout file (activity_main.xml) that defines the EditText and TextView elements, as well as the four buttons. You'll need to create this file and reference the appropriate IDs in your code.