Create CustomizeAlertDialog.kt
The size of an alert dialog can be customized using the AlertDialog class and its associated AlertDialog.Builder class. To make the dialog fill 90% of the screen size, you can set its dimensions programmatically using the WindowManager.LayoutParams class.
In the above code, the layoutParams.width property is set to 90% of the screen width, and the layoutParams.height property is set to WindowManager.LayoutParams.WRAP_CONTENT to automatically adjust the height based on the content of the dialog.
You can adjust the values as per your requirement to achieve the desired result.