Create VideoViewController.kt
This code initializes a VideoView component and sets up a MediaController for it. The MediaController provides a set of controls for playing and pausing the video, seeking forward and backward, and adjusting the volume.
The setAnchorView method is called to set the view that the media controller will be anchored to, which in this case is the VideoView.
The setVideoURI method is used to set the source of the video, which is a video file located in the raw resources folder of the app. The file is accessed using a Uri that is constructed by concatenating the package name of the app and the resource identifier of the video file.
Overall, this code sets up a VideoView component and a MediaController for it, and associates a video file as the source for the VideoView.