Create CustomAdapter.kt
If you required the custom adapter for recycler view.
This code defines an adapter class named "MyAdapter" for a RecyclerView. The adapter is responsible for providing the data to the RecyclerView and creating and managing its views.
The adapter class has a private member variable "mData" which is a List of Strings representing the data to be displayed in the RecyclerView.
The constructor of the adapter takes a List of Strings as a parameter and initializes the "mData" member variable.
The onCreateViewHolder method is responsible for creating a new ViewHolder for the RecyclerView. It inflates a layout file called "item_layout" and returns a new ViewHolder object that holds the inflated view.
The onBindViewHolder method is responsible for updating the contents of the ViewHolder with the data at the given position. In this case, it sets the text of the TextView in the ViewHolder to the data at the given position in the "mData" list.
The getItemCount method returns the number of items in the "mData" list.
The ViewHolder class is a static inner class that extends RecyclerView.ViewHolder. It holds a reference to the TextView view that is used to display the data in the RecyclerView. In its constructor, it initializes the TextView member variable by finding the view