Create NetworkSetupWithRetrofit.kt
The given code is an example of making a network request to the GitHub API using Retrofit, which is a popular HTTP client library for Android development.
In this code, a Retrofit instance is created with a baseUrl of "https://api.github.com/" and a GsonConverterFactory for converting JSON responses to objects. An OkHttpClient instance is also created with specific settings for the connection and read timeouts, as well as an HttpLoggingInterceptor for logging requests and responses. This OkHttpClient instance is then passed to the Retrofit.Builder using the client() method.
An interface named ApiService is defined with a method named getUsers() annotated with @GET("users") to specify the endpoint for the GET request. The method returns a List<User> object, which represents a list of users obtained from the API.
Finally, an instance of the ApiService interface is created using the Retrofit.create() method, and the getUsers() method is called to retrieve the list of users from the GitHub API.